Refactor collection argument parsing
