+++ aquadoggo/src/graphql/utils.rs
@@ -1,16 +1,24 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use async_graphql::dynamic::{ResolverContext, TypeRef, ValueAccessor};
+use std::convert::TryFrom;
+use std::num::NonZeroU64;
+
+use async_graphql::dynamic::{ResolverContext, TypeRef, ValueAccessor, ObjectAccessor};
 use async_graphql::{Error, Value};
+use dynamic_graphql::ScalarValue;
 use p2panda_rs::document::{DocumentId, DocumentViewId};
-use p2panda_rs::operation::{OperationValue, Relation};
-use p2panda_rs::schema::{FieldType, SchemaId};
+use p2panda_rs::operation::OperationValue;
+use p2panda_rs::schema::{FieldType, SchemaId, Schema};
 use p2panda_rs::storage_provider::error::DocumentStorageError;
 use p2panda_rs::storage_provider::traits::DocumentStore;
 
+use crate::db::query::{Filter, Order, Pagination, MetaField, Field, Direction};
 use crate::db::{types::StorageDocument, SqlStore};
 use crate::graphql::scalars::{DocumentIdScalar, DocumentViewIdScalar};
 use crate::graphql::types::DocumentValue;
+
+use super::constants;
+use super::scalars::CursorScalar;
 
 const DOCUMENT_FIELDS_SUFFIX: &str = "Fields";
 const FILTER_INPUT_SUFFIX: &str = "Filter";
@@ -109,6 +117,172 @@ pub fn filter_to_operation_value(
     };
 
     Ok(value)
+}
+
+/// Parse all argument values based on expected keys and types.
+pub fn parse_collection_arguments(
+    ctx: &ResolverContext,
+    schema: &Schema, 
+    pagination: &mut Pagination<CursorScalar>,
+    order: &mut Order,
+    meta_filter: &mut Filter,
+    field_filter: &mut Filter,
+) -> Result<(), Error> {
+    for (name, value) in ctx.args.iter() {
+        match name.as_str() {
+            constants::PAGINATION_CURSOR_ARG => {
+                let cursor = CursorScalar::from_value(Value::String(value.string()?.to_string()))?;
+                pagination.after = Some(cursor);
+            }
+            constants::PAGINATION_FIRST_ARG => {
+                pagination.first = NonZeroU64::try_from(value.u64()?)?;
+            }
+            constants::ORDER_BY_ARG => {
+                let order_by = match value.enum_name()? {
+                    "OWNER" => Field::Meta(MetaField::Owner),
+                    field_name => Field::new(field_name),
+                };
+                order.field = order_by;
+            }
+            constants::ORDER_DIRECTION_ARG => {
+                let direction = match value.enum_name()? {
+                    "ASC" => Direction::Ascending,
+                    "DESC" => Direction::Descending,
+                    _ => panic!("Unknown order direction argument key received"),
+                };
+                order.direction = direction;
+            }
+            constants::META_FILTER_ARG => {
+                let filter_object = value
+                    .object()
+                    .map_err(|_| Error::new("internal: is not an object"))?;
+                parse_meta_filter(meta_filter, &filter_object)?;
+            }
+            constants::FILTER_ARG => {
+                let filter_object = value
+                    .object()
+                    .map_err(|_| Error::new("internal: is not an object"))?;
+                parse_filter(field_filter, &schema, &filter_object)?;
+            }
+            _ => panic!("Unknown argument key received"),
+        }
+    }
+    Ok(())
+}
+
+
+/// Parse a filter object received from the graphql api into an abstract filter type based on the
+/// schema of the documents being queried.
+fn parse_filter(
+    filter: &mut Filter,
+    schema: &Schema,
+    filter_object: &ObjectAccessor,
+) -> Result<(), Error> {
+    for (field, filters) in filter_object.iter() {
+        let filter_field = Field::new(field.as_str());
+        let filters = filters.object()?;
+        for (name, value) in filters.iter() {
+            let field_type = schema.fields().get(field.as_str()).unwrap();
+            match name.as_str() {
+                "in" => {
+                    let mut list_items: Vec<OperationValue> = vec![];
+                    for value in value.list()?.iter() {
+                        let item = filter_to_operation_value(&value, field_type)?;
+                        list_items.push(item);
+                    }
+                    filter.add_in(&filter_field, &list_items);
+                }
+                "notIn" => {
+                    let mut list_items: Vec<OperationValue> = vec![];
+                    for value in value.list()?.iter() {
+                        let item = filter_to_operation_value(&value, field_type)?;
+                        list_items.push(item);
+                    }
+                    filter.add_not_in(&filter_field, &list_items);
+                }
+                "eq" => {
+                    let value = filter_to_operation_value(&value, field_type)?;
+                    filter.add(&filter_field, &value);
+                }
+                "notEq" => {
+                    let value = filter_to_operation_value(&value, field_type)?;
+                    filter.add_not(&filter_field, &value);
+                }
+                "gt" => {
+                    let value = filter_to_operation_value(&value, field_type)?;
+                    filter.add_gt(&filter_field, &value);
+                }
+                "gte" => {
+                    let value = filter_to_operation_value(&value, field_type)?;
+                    filter.add_gte(&filter_field, &value);
+                }
+                "lt" => {
+                    let value = filter_to_operation_value(&value, field_type)?;
+                    filter.add_lt(&filter_field, &value);
+                }
+                "lte" => {
+                    let value = filter_to_operation_value(&value, field_type)?;
+                    filter.add_lte(&filter_field, &value);
+                }
+                "contains" => {
+                    filter.add_contains(&filter_field, &value.string()?);
+                }
+                "notContains" => {
+                    filter.add_contains(&filter_field, &value.string()?);
+                }
+                _ => panic!("Unknown filter type received"),
+            }
+        }
+    }
+    Ok(())
+}
+
+/// Parse a meta filter object received from the graphql api into an abstract filter type based on the
+/// schema of the documents being queried.
+fn parse_meta_filter(filter: &mut Filter, filter_object: &ObjectAccessor) -> Result<(), Error> {
+    for (field, filters) in filter_object.iter() {
+        let meta_field = MetaField::try_from(field.as_str())?;
+        let filter_field = Field::Meta(meta_field);
+        let filters = filters.object()?;
+        for (name, value) in filters.iter() {
+            match name.as_str() {
+                "in" => {
+                    let mut list_items: Vec<OperationValue> = vec![];
+                    for value in value.list()?.iter() {
+                        let item = filter_to_operation_value(&value, &FieldType::String)?;
+                        list_items.push(item);
+                    }
+                    filter.add_in(&filter_field, &list_items);
+                }
+                "notIn" => {
+                    let mut list_items: Vec<OperationValue> = vec![];
+                    for value in value.list()?.iter() {
+                        let item = filter_to_operation_value(&value, &FieldType::String)?;
+                        list_items.push(item);
+                    }
+                    filter.add_not_in(&filter_field, &list_items);
+                }
+                "eq" => {
+                    let field_type = match field.as_str() {
+                        "edited" | "deleted" => FieldType::Boolean,
+                        _ => FieldType::String,
+                    };
+                    let value = filter_to_operation_value(&value, &field_type)?;
+                    filter.add(&filter_field, &value);
+                }
+                "notEq" => {
+                    let field_type = match field.as_str() {
+                        "edited" | "deleted" => FieldType::Boolean,
+                        _ => FieldType::String,
+                    };
+                    let value = filter_to_operation_value(&value, &field_type)?;
+                    filter.add_not(&filter_field, &value);
+                }
+                _ => panic!("Unknown meta filter type received"),
+            }
+        }
+    }
+    Ok(())
 }
 
 /// Downcast document value which will have been passed up by the parent query node,
